Extracted area detector tests into their own class
